remove unused code and imports refactor navigation to use nextController annotation
